refactor: #16 도메인별 repository 패키지 경로 개선
