--- src/main/java/com/wholeseeds/mindle/domain/complaint/repository/CategoryRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.wholeseeds.mindle.domain.complaint.repository;
-
-public interface CategoryRepositoryCustom {
-}
