--- src/main/java/com/wholeseeds/mindle/domain/complaint/repository/ComplaintRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.wholeseeds.mindle.domain.complaint.repository;
-
-public interface ComplaintRepositoryCustom {
-}
