+++ src/main/java/com/wholeseeds/mindle/domain/complaint/repository/custom/CategoryRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.wholeseeds.mindle.domain.complaint.repository.custom;
+
+public interface CategoryRepositoryCustom {
+}
