+++ src/main/java/com/wholeseeds/mindle/domain/complaint/repository/custom/ComplaintImageRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.wholeseeds.mindle.domain.complaint.repository.custom;
+
+public interface ComplaintImageRepositoryCustom {
+}
