+++ src/main/java/com/wholeseeds/mindle/domain/complaint/repository/impl/CategoryRepositoryImpl.java
@@ -1,8 +1,9 @@
-package com.wholeseeds.mindle.domain.complaint.repository;
+package com.wholeseeds.mindle.domain.complaint.repository.impl;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepositoryImpl;
 import com.wholeseeds.mindle.domain.complaint.entity.Category;
 import com.wholeseeds.mindle.domain.complaint.entity.QCategory;
+import com.wholeseeds.mindle.domain.complaint.repository.custom.ComplaintRepositoryCustom;
 
 import jakarta.persistence.EntityManager;
