+++ src/main/java/com/wholeseeds/mindle/domain/complaint/repository/impl/ComplaintRepositoryImpl.java
@@ -1,8 +1,9 @@
-package com.wholeseeds.mindle.domain.complaint.repository;
+package com.wholeseeds.mindle.domain.complaint.repository.impl;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepositoryImpl;
 import com.wholeseeds.mindle.domain.complaint.entity.Complaint;
 import com.wholeseeds.mindle.domain.complaint.entity.QComplaint;
+import com.wholeseeds.mindle.domain.complaint.repository.custom.ComplaintRepositoryCustom;
 
 import jakarta.persistence.EntityManager;
