--- src/main/java/com/wholeseeds/mindle/domain/location/repository/CityRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.wholeseeds.mindle.domain.location.repository;
-
-public interface CityRepositoryCustom {
-}
