--- src/main/java/com/wholeseeds/mindle/domain/location/repository/SubdistrictRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.wholeseeds.mindle.domain.location.repository;
-
-public interface SubdistrictRepositoryCustom {
-}
