+++ src/main/java/com/wholeseeds/mindle/domain/location/repository/custom/CityRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.wholeseeds.mindle.domain.location.repository.custom;
+
+public interface CityRepositoryCustom {
+}
