+++ src/main/java/com/wholeseeds/mindle/domain/location/repository/impl/DistrictRepositoryImpl.java
@@ -1,8 +1,9 @@
-package com.wholeseeds.mindle.domain.location.repository;
+package com.wholeseeds.mindle.domain.location.repository.impl;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepositoryImpl;
 import com.wholeseeds.mindle.domain.location.entity.District;
 import com.wholeseeds.mindle.domain.location.entity.QDistrict;
+import com.wholeseeds.mindle.domain.location.repository.custom.DistrictRepositoryCustom;
 
 import jakarta.persistence.EntityManager;
