+++ src/main/java/com/wholeseeds/mindle/domain/member/repository/MemberRepository.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Repository;
 
 import com.wholeseeds.mindle.common.repository.JpaBaseRepository;
 import com.wholeseeds.mindle.domain.member.entity.Member;
+import com.wholeseeds.mindle.domain.member.repository.custom.MemberRepositoryCustom;
 
 @Repository
 public interface MemberRepository extends JpaBaseRepository<Member, Long>, MemberRepositoryCustom {
