--- src/main/java/com/wholeseeds/mindle/domain/place/repository/PlaceRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.wholeseeds.mindle.domain.place.repository;
-
-public interface PlaceRepositoryCustom {
-}
