+++ src/main/java/com/wholeseeds/mindle/domain/place/repository/custom/PlaceRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.wholeseeds.mindle.domain.place.repository.custom;
+
+public interface PlaceRepositoryCustom {
+}
